% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airqES.R
\docType{data}
\name{airqES}
\alias{airqES}
\title{Air quality measurements in Spain}
\format{
A \code{data.frame} with 112,484 rows and columns:
\itemize{
\item year, Year of measurement
\item month, Month of measurement
\item province, Province
\item city, City
\item station_id, Measurement station identifier
\item pollutant, Pollutant name
\item D01-D031, days 1-31 of measurement
  }
}
\source{
Official data  \href{www.miteco.gob.es}{Government of Spain}.
}
\usage{
data(airqES)
}
\description{
Daily data referred to several pollutants (from 2001 to 2018).
}
\author{
Jose V. Die & Jose R. Caro
}
\keyword{datasets}
