% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquaityES.R
\docType{package}
\name{airqualityES}
\alias{airqualityES}
\alias{airqualityES-package}
\title{airqualityES: Air quality measurements in Spain}
\description{
aiqualityES datasets contain daily quality air measurements from 2001 to 2018.
}
\seealso{
see the preprint in \href{https://www.biorxiv.org/content/10.1101/436659v1}{BioRvix}
}
\author{
Jose V. Die \email{jose.die@uco.es}, Jose R. Caro
}
