% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Air stations in Spain}
\format{A \code{data.frame} with 604 rows and columns:
\itemize{
\item id, local identifier
\item province_id, province identifier
\item city_id, city identifier
\item station, station identifier
\item location_id, location identifier
\item administration, official administration
\item station_name, name of the station
\item first_day, day of first recod
\item last_day, day of last record
\item latitude, coordinates
\item longitude, coordinates
\item altitude, coordinates
\item auton_commun, autonomous communities
\item province, province
\item city, city
\item station_kind, kind of station
\item area, area
  }}
\source{
Official data  \href{www.miteco.gob.es}{Government of Spain}.
}
\usage{
data(stations)
}
\description{
Description of station features.
}
\author{
Jose V. Die & Jose R. Caro
}
\keyword{datasets}
