% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquaityES.R
\docType{package}
\name{airqualityES}
\alias{airqualityES}
\alias{airqualityES-package}
\title{airqualityES: Air quality measurements in Spain}
\description{
aiqualityES dataset contains daily quality air measurements from 2001 to 2018.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jdieramon/airqualityES}
  \item Report bugs at \url{https://github.com/jdieramon/airqualityES/issues}
}

}
\author{
Jose V. Die \email{jose.die@uco.es}, Jose R. Caro
}
